requirements.txt
quart
tortoise-orm
pydantic[dotenv]
